backend/requirements.txt
celery
celery[redis]
Django
django-cors-headers
django-filter
djangorestframework
drf-spectacular
parametrize
pyjwt[crypto]
pyyaml
uritemplate
